// core/src/core/model/TokenType.ts
/**
 * All token types defined in the DTCG 2025.10 specification.
 *
 * @see https://tr.designtokens.org/format/#types
 */
export type TokenType =
    | "color"
    | "dimension"
    | "fontFamily"
    | "fontWeight"
    | "number"
    | "duration"
    | "cubicBezier"
    | "strokeStyle"
    | "border"
    | "transition"
    | "shadow"
    | "gradient"
    | "typography";